[#1014] - Elimina tipos derivados y reimplementa los generados por Sanity TypeGen
