+++ src/api/_queries/storylist.query.ts
@@ -1,6 +1,5 @@
 import { defineQuery } from 'groq';
 
-// @sanity-typegen-ignore
 export const storylistTeasersQuery = defineQuery(`*[_type == 'storylist']{ 
     'slug': slug.current,
     title,
@@ -21,7 +20,6 @@ export const storylistTeasersQuery = defineQuery(`*[_type == 'storylist']{
     }
 `);
 
-// @sanity-typegen-ignore
 export const storylistQuery = defineQuery(`*[_type == 'storylist' && slug.current == $slug][0]
 { 
     'slug': slug.current,
@@ -48,11 +46,19 @@ export const storylistQuery = defineQuery(`*[_type == 'storylist' && slug.curren
             title,
             language,
             badLanguage,
-            body[0...3],
+            'body': coalesce(body[0...3], []),
             originalPublication,
             approximateReadingTime,
+            'resources': [],
             'mediaSources': coalesce(mediaSources[], []),
-            'author': author-> { slug, name, image, nationality->, 'biography': [] }
+            'author': author->{ 
+                slug,
+                name,
+                image,
+                nationality->,
+                'biography': [],
+                'resources': [],
+            }
         }
     }, []),
     'count': count(*[ _type == 'publication' && storylist._ref == ^._id ])
